Refactor: widgets module
